sdl2: updated to git 3d22731 / disable SSEMATH
- https://github.com/libsdl-org/SDL/blob/release-2.0.14/CMakeLists.txt#L656-L666

‘387’
Use the standard 387 floating-point coprocessor present on the majority of chips and emulated otherwise. Code compiled with this option runs almost everywhere. The temporary results are computed in 80-bit precision instead of the precision specified by the type, resulting in slightly different results compared to most of other chips. See -ffloat-store for more detailed description.

This is the default choice for non-Darwin x86-32 targets.

- https://gcc.gnu.org/onlinedocs/gcc/x86-Options.html